+++ configure.py
@@ -493,6 +493,7 @@ for name in ['build',
              'manifest_parser',
              'metrics',
              'state',
+             'status',
              'string_piece_util',
              'util',
              'version']:
@@ -556,6 +557,7 @@ for name in ['build_log_test',
              'manifest_parser_test',
              'ninja_test',
              'state_test',
+             'status_test',
              'string_piece_util_test',
              'subprocess_test',
              'test',
